General cleanup and variable streamlining now using primarily string and doubles.

+++ CopilotApp/CopilotApp/CopilotApp/LiveData/SensorDataBackingFields.cs
@@ -9,25 +9,25 @@ namespace CopilotApp
         //Front Left Sensor Backing Fields
         public static string _frontLeftSensorID;
         public static string _frontLeftSensorStatus;
-        public static int _frontLeftSensorTemperature;
+        public static double _frontLeftSensorTemperature;
         public static double _frontLeftSensorPressure;
 
         //Front Right Sensor Backing Fields
         public static string _frontRightSensorID;
         public static string _frontRightSensorStatus;
-        public static int _frontRightSensorTemperature;
+        public static double _frontRightSensorTemperature;
         public static double _frontRightSensorPressure;
 
         //Rear Left Sensor Backing Fields
         public static string _rearLeftSensorID;
         public static string _rearLeftSensorStatus;
-        public static int _rearLeftSensorTemperature;
+        public static double _rearLeftSensorTemperature;
         public static double _rearLeftSensorPressure;
 
         //Rear Right Sensor Backing Fields
         public static string _rearRightSensorID;
         public static string _rearRightSensorStatus;
-        public static int _rearRightSensorTemperature;
+        public static double _rearRightSensorTemperature;
         public static double _rearRightSensorPressure;
     }
 }
